ESP32: Fixed the type of cpuint variables in esp32_emac.c esp32_i2c.c esp32_spi.c esp32_spi_slave.c

+++ arch/xtensa/src/esp32/esp32_gpio.c
@@ -57,7 +57,7 @@
  ****************************************************************************/
 
 #ifdef CONFIG_ESP32_GPIO_IRQ
-static uint8_t g_gpio_cpuint;
+static int g_gpio_cpuint;
 #endif
 
 static const uint8_t g_pin2func[40] =
